Unification of links & icons
