--- src/meta/icons/index.ts
@@ -1,7 +1,0 @@
-export interface IconMeta {
-  color?: string
-  fill?: boolean
-  extraBc?: string
-  extraBcRad?: string
-  svg: any
-}
